+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-### Fixed
+### Added
 * Support Rails 8.1, Ruby 4.0
 * capistrano: Support bundler 4 with capistrano 2