+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
-*no unreleased changes*
+### Added
+* Support Rails 8.1, Ruby 4.0
+* capistrano: Support bundler 4 with capistrano 2
 
 ## 7.3.5 / 2025-10-16
 ### Fixed